Improved automatic patch handling
Switched from "git apply" to "patch". No longer
relies on git for applying patches.

The switch also solves the problems of "git apply"
only working in the repo root, and needing git
installed to install or uninstall patches.

This will make it easier in situations where
installing git may be extra overhead, and free
the user from having to use specific working
directories for the script.