Add option to automatically jump to applications
In many cases you don't need to setup or configure anything,
you just want to automatically start or get to the
application selection screen.

This adds the configuration option to automatically go to
the applications selection window. It also triggers the
startup of applications that were configured to auto startup.