Add selective auto-close of applications
Supports the ability to have an application that doesn't
auto-close when exiting the application run menu.

This means that only some applications will be shutdown and
restarted when using the auto-start and close menu options.

Will also detect when an application is already running
and not start that application up a second time.